requirements.txt
# Used to package the debian package
# Pins versions to ensure the package is built consistently and reproducibly

flask==2.3.2
flask-events==0.7.0
mohawk==1.0.0
pyyaml==5.4
requests==2.21.0
gevent==1.4.0
gunicorn==19.9.0

# gevent deps
greenlet==0.4.15

# flask deps
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==0.15.5
click==7.0
itsdangerous==1.1.0

# mohawk deps
six==1.12.0

# requests deps
certifi==2022.12.7
chardet==3.0.4
idna==2.8
urllib3==1.26.5
